feat: Update ActionType to be an array of strings in ActionLog interface
The ActionType property in the ActionLog interface has been updated to be an array of strings instead of a single string. This change allows for multiple action types to be associated with an action log entry. The code changes include modifying the ActionType property type in the ContentContext.tsx file.